Working on Back-end API

+++ src/components/category-element/CategoryElement.tsx
@@ -1,3 +1,5 @@
+"use client"
+
 import style from "./CategoryElement.module.css"
 import Image from "next/image"
 import trashIcon from "@/public/trash_icon.png"
@@ -7,15 +9,28 @@ import { DBCategory } from "@/src/types/Categories"
 import { deleteCategoryAction } from "@/src/lib/actions/categoryAction"
 import { useActionState } from "react"
 import { HttpStatus } from "@/src/enums"
+import { useConfirm } from "../basic/popups/confirm-popup/ConfirmPopup"
 
 
 export default function CategoryElement({category, deleteCallback} : {category: DBCategory, deleteCallback: Function}) {
-    const deleteAction = async (prevState: any, formData: FormData) => {
-        const response = await deleteCategoryAction(prevState, formData);
-        if (!response.successful) alert(response.message);
+    const confirm  = useConfirm();
 
-        const id = response.data != null ? response.data.id : null;
-        deleteCallback(id);
+    const deleteAction = async (prevState: any, formData: FormData) => {
+        let response = null;
+
+        setTimeout(async() => {
+            const ok = await confirm({title: "Deleting...", message: "Are you sure you want to delete this category?"})
+            
+            if (!ok) return;
+
+            response = await deleteCategoryAction(prevState, formData);
+            if (!response.successful) alert(response.message);
+
+            const id = response.data != null ? response.data.id : null;
+            deleteCallback(id);
+            
+        }, 0);
+
         return response
     }
 
@@ -24,7 +39,7 @@ export default function CategoryElement({category, deleteCallback} : {category: 
    return (
     <form className={style.category} action={formAction}>
         <input type="hidden" name="categoryId" value={category.id} />
-        <Button>
+        <Button attributes={{className: style.openButton, type: "button"}}>
             <Image src={categoryIconsPath + category.iconName} alt={""} width={20} height={20}></Image>
             {category.name}
         </Button>
